backend/requirements.txt
fastapi==0.115.12
uvicorn==0.34.0
pydantic==2.11.2
python-dateutil==2.9.0.post0
python-jose==3.4.0
bcrypt>=4.0.0
python-dotenv==1.0.1
psycopg[binary]==3.2.9
